+++ src/test/java/org/jabref/logic/formatter/bibtexfields/ClearFormatterTest.java
@@ -8,12 +8,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 /**
  * Tests in addition to the general tests from {@link org.jabref.logic.formatter.FormatterTest}
  */
-public class ClearFormatterTest {
+class ClearFormatterTest {
 
     private ClearFormatter formatter;
 
     @BeforeEach
-    public void setUp() {
+    void setUp() {
         formatter = new ClearFormatter();
     }
 
@@ -21,7 +21,7 @@ public class ClearFormatterTest {
      * Check whether the clear formatter really returns the empty string for the empty string
      */
     @Test
-    public void formatReturnsEmptyForEmptyString() throws Exception {
+    void formatReturnsEmptyForEmptyString() throws Exception {
         assertEquals("", formatter.format(""));
     }
 
@@ -29,12 +29,12 @@ public class ClearFormatterTest {
      * Check whether the clear formatter really returns the empty string for some string
      */
     @Test
-    public void formatReturnsEmptyForSomeString() throws Exception {
+    void formatReturnsEmptyForSomeString() throws Exception {
         assertEquals("", formatter.format("test"));
     }
 
     @Test
-    public void formatExample() {
+    void formatExample() {
         assertEquals("", formatter.format(formatter.getExampleInput()));
     }
 }