+++ src/test/java/org/jabref/logic/formatter/casechanger/ProtectTermsFormatterTest.java
@@ -13,44 +13,44 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 /**
  * Tests in addition to the general tests from {@link org.jabref.logic.formatter.FormatterTest}
  */
-public class ProtectTermsFormatterTest {
+class ProtectTermsFormatterTest {
 
     private ProtectTermsFormatter formatter;
 
     @BeforeEach
-    public void setUp() {
+    void setUp() {
         formatter = new ProtectTermsFormatter(
                 new ProtectedTermsLoader(new ProtectedTermsPreferences(ProtectedTermsLoader.getInternalLists(),
                         Collections.emptyList(), Collections.emptyList(), Collections.emptyList())));
     }
 
     @Test
-    public void singleWord() {
+    void singleWord() {
         assertEquals("{VLSI}", formatter.format("VLSI"));
     }
 
     @Test
-    public void doNotProtectAlreadyProtected() {
+    void doNotProtectAlreadyProtected() {
         assertEquals("{VLSI}", formatter.format("{VLSI}"));
     }
 
     @Test
-    public void caseSensitivity() {
+    void caseSensitivity() {
         assertEquals("VLsI", formatter.format("VLsI"));
     }
 
     @Test
-    public void formatExample() {
+    void formatExample() {
         assertEquals("In {CDMA}", formatter.format(formatter.getExampleInput()));
     }
 
     @Test
-    public void correctOrderingOfTerms() {
+    void correctOrderingOfTerms() {
         assertEquals("{3GPP} {3G}", formatter.format("3GPP 3G"));
     }
 
     @Test
-    public void test() {
+    void test() {
         assertEquals("{VLSI} {VLSI}", formatter.format("VLSI {VLSI}"));
         assertEquals("{BPEL}", formatter.format("{BPEL}"));
         assertEquals("{Testing {BPEL} Engine Performance: A Survey}",