+++ src/test/java/org/jabref/logic/importer/fileformat/ACMPortalParserTest.java
@@ -24,7 +24,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 @FetcherTest
-public class ACMPortalParserTest {
+class ACMPortalParserTest {
 
     ACMPortalParser parser;
     List<BibEntry> searchEntryList;