+++ src/test/java/org/jabref/logic/integrity/CitationKeyDuplicationCheckerTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class CitationKeyDuplicationCheckerTest {
+class CitationKeyDuplicationCheckerTest {
 
     @Test
     void emptyCitationKey() {