+++ src/test/java/org/jabref/logic/layout/format/AuthorFirstAbbrLastOxfordCommasTest.java
@@ -6,13 +6,13 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class AuthorFirstAbbrLastOxfordCommasTest {
+class AuthorFirstAbbrLastOxfordCommasTest {
 
     /**
      * Test method for {@link org.jabref.logic.layout.format.AuthorFirstAbbrLastOxfordCommas#format(java.lang.String)}.
      */
     @Test
-    public void format() {
+    void format() {
         LayoutFormatter a = new AuthorFirstAbbrLastOxfordCommas();
 
         // Empty case