+++ src/test/java/org/jabref/logic/util/strings/StringLengthComparatorTest.java
@@ -9,12 +9,12 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class StringLengthComparatorTest {
+class StringLengthComparatorTest {
 
     private StringLengthComparator slc;
 
     @BeforeEach
-    public void setUp() {
+    void setUp() {
         slc = new StringLengthComparator();
     }